refactor(frontend): migrate dashboard to feature-based architecture
