+++ frontend/src/features/dashboard/components/leaf/NavItem.tsx
@@ -1,5 +1,5 @@
-import { useTheme } from '../../hooks/useTheme'
-import { cn } from '../../lib/utils'
+import { useTheme } from '../../../../hooks/useTheme'
+import { cn } from '../../../../lib/utils'
 
 type NavItemProps = {
   icon: string
