+++ frontend/src/features/dashboard/components/leaf/ThemeToggle.tsx
@@ -1,6 +1,6 @@
 import { Sun, Moon } from 'lucide-react'
-import { useTheme } from '../../hooks/useTheme'
-import { cn } from '../../lib/utils'
+import { useTheme } from '../../../../hooks/useTheme'
+import { cn } from '../../../../lib/utils'
 
 /**
  * ThemeToggle button for switching between light and dark modes.
